+++ Cargo.toml
@@ -36,6 +36,8 @@ curve25519-dalek = { version = "4.0.0", features = ["digest", "rand_core", "serd
 rand_core = "0.6.4"
 sha2 = "0.10.8"
 pasta_curves = "0.5.1"
+bellman = "0.14.0"
+pairing = "0.23.0"
 
 [build-dependencies]
 tonic-build = "0.10.2"
